+++ database/migrations/2020_05_21_170659_create_sold_sales_rooms_table.php
@@ -16,10 +16,10 @@ class CreateSoldSalesRoomsTable extends Migration
         Schema::create('sold_sales_rooms', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedBigInteger('room_id')->index();
-            $table->integer('price')->default(0);//成約価格（万円）
-            $table->integer('previous_price')->default(0);//成約前価格（万円）
-            $table->integer('management_fee')->default(0);//管理費
-            $table->integer('reserve_fund')->default(0);//修繕積立金
+            $table->float('price',8,2)->default(0);//成約価格（万円）
+            $table->float('previous_price',8,2)->default(0);//成約前価格（万円）
+            $table->float('management_fee',8,2)->default(0);//管理費
+            $table->float('reserve_fund',8,2)->default(0);//修繕積立金
             $table->char('company_name',100)->nullable();//会員名
             $table->char('contact_phonenumber',100)->nullable();//代表電話番号
             $table->char('pic',100)->nullable();//問合せ担当者
